### 2023-05-08 7671bca
Revert "add the function for subsetting the staggered archive"
This reverts commit 449486f5465c4bd5580d452b081281592c8b2aa6.
