### 2023-05-08 4fd4fae
Revert "Revert "add the function for subsetting the staggered archive""
This reverts commit 7671bcabf48c56c8164fc5c0cc994fdb8f42bfae.
