### 2023-05-08 3edb56d
flag on nearest neighbor distances
