### 2023-05-15 768fa19
fix the not keyword on the flag and test it
