+++ requirements.txt
@@ -9,3 +9,4 @@ scikit-learn>=1.1
 gcsfs>=2022.5.0
 fsspec>=2022.5.0
 tqdm>=4.64.1
+urllib3<2.0
