### 2023-05-15 981c744
see i requiring urllib3<2.0 first will actually make GA do it
GA ignoring right now
