+++ requirements.txt
@@ -1,3 +1,4 @@
+urllib3<2.0
 matplotlib>=3.3.2
 xarray>=2022.9.0
 numpy>=1.23.3
@@ -9,4 +10,3 @@ scikit-learn>=1.1
 gcsfs>=2022.5.0
 fsspec>=2022.5.0
 tqdm>=4.64.1
-urllib3<2.0
