+++ requirements.txt
@@ -1,4 +1,4 @@
-urllib3<2.0
+urllib3<2.0, >=1.21.1
 matplotlib>=3.3.2
 xarray>=2022.9.0
 numpy>=1.23.3
