### 2023-06-07 #5525 @editorialbot
[REVIEW]: STITCHES: a Python package to amalgamate existing Earth system model output into new scenario realizations
**Submitting author:** @abigailsnyder (<a href="http://orcid.org/0000-0002-9034-9948">Abigail Snyder</a>)
**Repository:** https://github.com/JGCRI/stitches
**Branch with paper.md** (empty if default branch): main
**Version:** v0.13
