### 2023-07-07 JGCRI/stitches#76 @znicholls
JOSS review: installation
At the moment, the installation instructions are only for installation from source. It would be great to add the ability to install from pip (and conda if you're feeling ambitious) as it is the standard packaging solution for Python.

https://github.com/openjournals/joss-reviews/issues/5525
