### 2023-07-07 JGCRI/stitches#77 @znicholls
JOSS review: README
A few things in the README aren't super clear as to their meaning.

- What does "To install pre-built run the following" mean? "To install pre-built, example data run the following"?
- What does "For users that would like to generate the run the following" mean? "For users that would like to generate the full data, run the following?"
- `stitches.generate_pkg_data` seems to also not exist, has it been re-named?

https://github.com/openjournals/joss-reviews/issues/5525
