### 2023-07-07 JGCRI/stitches#78 @znicholls
JOSS review: notebooks
- https://github.com/JGCRI/stitches/blob/main/notebooks/stitches-quickstart.ipynb runs nicely, there are a few tweaks which could help I think:
  - explain in a bit more detail what exactly the `get_chunk_info` and `chunk_ts` functions are doing
  - show the dataframes that are being used (e.g. what does my recipes look like, archive data etc.)
  - explain how the archive data was produced in the first place (this seems fairly key and I think an example would really help)
  - variable names could be slightly clearer e.g. `data` doesn't really tell the user anything about what the variable is
  - the diagram is super nice, well done on that

- is https://github.com/JGCRI/stitches/blob/main/notebooks/full_pipeline_example.ipynb meant to be for users or is that just internal?

https://github.com/openjournals/joss-reviews/issues/5525
