### 2023-07-07 JGCRI/stitches#79 @znicholls
JOSS review: community guidelines
I think the formatting doesn't quite render correctly at https://jgcri.github.io/stitches/reference/contributing.html (some things appear in quotes when I don't think they're meant to)

Otherwise I thought these were ok, but @Zeitsperre makes lots of good suggestions in their review

https://github.com/openjournals/joss-reviews/issues/5525
