### 2023-07-07 JGCRI/stitches#80 @znicholls
JOSS review: tests
The tests ran for me although one was failing on main. The instructions are all clear, but using `python setup.py` is pretty old school now I think (I think there's been a big move to `pyproject.toml` in order to better support different possible build backends).

It would be helpful to document how to get coverage information for devs (although it is all based on standard tools so isn't too hard to work out).

46% test coverage isn't super high, it could be worth thinking about whether it is feasible to push that figure higher (or remove the less used/tested bits of the API).

https://github.com/openjournals/joss-reviews/issues/5525
