### 2023-07-07 JGCRI/stitches#79 @Zeitsperre
Thanks for opening the discussion, @znicholls. I'm mostly fine with the community guidelines, but there are a few low-hanging fruits that would make the project much better, IMO.

In addition to the rendering issue that Zeb mentioned, a few suggestions:
* There are both a `CONTRIBUTING.md` and a `contributing.rst` that contain similar information. A way of synchronizing them would be to use the MyST extnsion (https://cerodell.github.io/sphinx-quickstart-guide/build/html/markdown.html).
* Having a statement on how best to report bugs or significant problems would be great.

Incidentally, merging of the information from both of the `Contributing` files would address most of my concerns around development environment setup and most of my comments around this.

My last comment concerned the Contributor Covenant contact person of contact; I've seen other versions of the Community Guidelines that include this information and others that don't; Feel free to disregard that comment from me.
