### 2023-07-07 JGCRI/stitches#76 @Zeitsperre
I looked into this in my review and unfortunately, the `STITCHES` name is taken on PyPI, (so it might be worthwhile to rename this project?). Having the project pip-installable would also help in self-enforcing packaging/metadata standards-compliance, in addition to other added benefits like discoverability.

Inclusion of `stitches in conda-forge would be fantastic for many climate research applications, but perhaps further down-the-line. I would gladly offer a hand in that once the project is in a more production-ready state.
