### 2023-07-07 #5525 @observingClouds
@znicholls, regarding your potential COI, could you let me know in which aspect you fall under the [JOSS COI policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html?highlight=conflict#joss-conflict-of-interest-policy)? Thank you!
