### 2023-07-08 #5525 @znicholls
> could you let me know in which aspect you fall under the JOSS COI policy?

In which I aspect I fall under or fail under? I have published papers with both Kalyn and Claudia in the last two years. I am also on a scientific steering committee with Claudia and plan to publish with Kalyn again in the next 12 months.
