### 2023-07-10 df02618
update installation call and Actions versions
