+++ .github/workflows/workflow.yml
@@ -16,17 +16,18 @@ jobs:
 
       steps:
 
-        - uses: actions/checkout@v1
+        - uses: actions/checkout@v3
 
         - name: Set up Python
-          uses: actions/setup-python@master
+          uses: actions/setup-python@v4
           with:
             python-version: 3.9
 
         - name: Install dependencies
           run: |
             python -m pip install --upgrade pip
-            python setup.py install
+            python -m pip install -r requirements.txt
+            python -m pip install .
             python -c 'import stitches; stitches.install_package_data()'
         - name: Test and generate coverage report on Linux
           run: |
