### 2023-07-10 b7f64bf
string formatting and typo fixes
