+++ stitches/tests/test_stitch.py
@@ -1,63 +1,88 @@
-import numpy as np
-import pandas as pd
 import os
 import unittest
+
+import numpy as np
+import pandas as pd
 import xarray as xr
 
-from stitches.fx_util import nrow
-from stitches.fx_stitch import find_var_cols, find_zfiles, internal_stitch, gmat_stitching, gridded_stitching
 from stitches.fx_pangeo import fetch_nc
+from stitches.fx_stitch import (
+    find_var_cols,
+    find_zfiles,
+    gmat_stitching,
+    gridded_stitching,
+    internal_stitch,
+)
+from stitches.fx_util import nrow
 
 
 class TestStitch(unittest.TestCase):
-
     # flag for run all or for ci
     RUN = "ci"
 
     # This is an example recipe that will be used to test the stitching functions
-    MY_RP = pd.DataFrame(data={'target_start_yr': [1850, 1859],
-                               'target_end_yr': [1858, 1867],
-                               'archive_experiment': ["historical", "historical"],
-                               'archive_variable': ["tas", "tas"],
-                               'archive_model': ["BCC-CSM2-MR", "BCC-CSM2-MR"],
-                               'archive_ensemble': ["r1i1p1f1", "r1i1p1f1"],
-                               'stitching_id': ["ssp245~r1i1p1f1~1", "ssp245~r1i1p1f1~1"],
-                               'archive_start_yr': [1859, 1886],
-                               'archive_end_yr': [1867, 1894],
-                               'tas_file': ["gs://cmip6/CMIP6/CMIP/BCC/BCC-CSM2-MR/historical/r1i1p1f1/Amon/tas/gn/v20181126/",
-                                           "gs://cmip6/CMIP6/CMIP/BCC/BCC-CSM2-MR/historical/r1i1p1f1/Amon/tas/gn/v20181126/"]})
+    MY_RP = pd.DataFrame(
+        data={
+            "target_start_yr": [1850, 1859],
+            "target_end_yr": [1858, 1867],
+            "archive_experiment": ["historical", "historical"],
+            "archive_variable": ["tas", "tas"],
+            "archive_model": ["BCC-CSM2-MR", "BCC-CSM2-MR"],
+            "archive_ensemble": ["r1i1p1f1", "r1i1p1f1"],
+            "stitching_id": ["ssp245~r1i1p1f1~1", "ssp245~r1i1p1f1~1"],
+            "archive_start_yr": [1859, 1886],
+            "archive_end_yr": [1867, 1894],
+            "tas_file": [
+                "gs://cmip6/CMIP6/CMIP/BCC/BCC-CSM2-MR/historical/r1i1p1f1/Amon/tas/gn/v20181126/",
+                "gs://cmip6/CMIP6/CMIP/BCC/BCC-CSM2-MR/historical/r1i1p1f1/Amon/tas/gn/v20181126/",
+            ],
+        }
+    )
 
     def test_find_var_cols(self):
-        o = pd.DataFrame(data={'tas': [1, 2], 'col2': [3, 4]})
+        o = pd.DataFrame(data={"tas": [1, 2], "col2": [3, 4]})
         self.assertEqual(len(find_var_cols(o)), 0)
 
-        o = pd.DataFrame(data={'tas_file': [1, 2], 'col2': [3, 4]})
-        self.assertEqual(find_var_cols(o), ['tas'])
+        o = pd.DataFrame(data={"tas_file": [1, 2], "col2": [3, 4]})
+        self.assertEqual(find_var_cols(o), ["tas"])
 
-        o = pd.DataFrame(data={'tas_file': [1, 2], 'col2': [3, 4], 'fake_file': [1, 2]})
[…]
