+++ stitches/__init__.py
@@ -1,12 +1,11 @@
 from .fx_match import match_neighborhood
-from .fx_recipe import permute_stitching_recipes, generate_gridded_recipe, make_recipe
-from .fx_stitch import gridded_stitching, gmat_stitching
-from .fx_pangeo import fetch_pangeo_table, fetch_nc
-from .make_tas_archive import make_tas_archive
-from .make_matching_archive import make_matching_archive
-from .make_pangeo_table import make_pangeo_table, make_pangeo_comparison
+from .fx_pangeo import fetch_nc, fetch_pangeo_table
+from .fx_recipe import generate_gridded_recipe, make_recipe, permute_stitching_recipes
+from .fx_stitch import gmat_stitching, gridded_stitching
 from .install_pkgdata import install_package_data
+from .make_matching_archive import make_matching_archive
+from .make_pangeo_table import make_pangeo_comparison, make_pangeo_table
+from .make_tas_archive import make_tas_archive
 from .package_data import *
-
 
 __version__ = "0.10.0"
