+++ stitches/fx_processing.py
@@ -11,21 +11,23 @@ import stitches.fx_util as util
 
 
 def calculate_rolling_mean(data, size):
-    """"
-  Calculate the rolling mean for the data frame with a user defined size centered window.
-  :param data:        A data frame of the cmip absolute temperature
-  :type data:        pandas.core.frame.DataFrame
-  :param size:          An integer value for the size of the window to use when calculating the rolling mean
-  :type size:           int
-  :return:          A pandas data frame of the smoothed time series (rolling mean applied)
-  """
+    """ "
+    Calculate the rolling mean for the data frame with a user defined size centered window.
+    :param data:        A data frame of the cmip absolute temperature
+    :type data:        pandas.core.frame.DataFrame
+    :param size:          An integer value for the size of the window to use when calculating the rolling mean
+    :type size:           int
+    :return:          A pandas data frame of the smoothed time series (rolling mean applied)
+    """
     # Check inputs
-    util.check_columns(data, {'model', 'experiment', 'ensemble', 'year', 'value', 'variable'})
+    util.check_columns(
+        data, {"model", "experiment", "ensemble", "year", "value", "variable"}
+    )
 
     # Index the data frame by the year, so that the rolling mean respects the years
-    data.index = data['year']
+    data.index = data["year"]
 
-    group_by = ['model', 'experiment', 'ensemble', 'variable']
+    group_by = ["model", "experiment", "ensemble", "variable"]
 
     # previously, we just had a call:
     # rslt = data.groupby(group_by)['value'].rolling(size, center=True).mean().reset_index()
@@ -42,40 +44,44 @@ def calculate_rolling_mean(data, size):
     # in a call to `transform` and save that as a new column:
 
     rslt = data.copy()
-    rslt['rollingAvg'] = rslt.groupby(group_by)['value'].transform(
-        lambda x: x.rolling(size, center=True, min_periods=1).mean())
+    rslt["rollingAvg"] = rslt.groupby(group_by)["value"].transform(
+        lambda x: x.rolling(size, center=True, min_periods=1).mean()
+    )
     # ^ seems like the way people usually get the functionality of a
     # dplyr group_by %>% mutate()  call.
 
     # rename so that value is the smoothed data:
-    rslt = rslt.drop(columns='value', axis=1).rename(columns={'rollingAvg':'value'}).reset_index(drop=True)
+    rslt = (
+        rslt.drop(columns="value", axis=1)
+        .rename(columns={"rollingAvg": "value"})
+        .reset_index(drop=True)
+    )
 
     return rslt
 
 
 def chunk_ts(df, n, base_chunk=0):
     """Format a data frame into an array of data frames containing data for n-sized years of successive data.
-  :param df:     data frame of climate data to chunk into different periods
-  :type df:     pandas DataFrame
-  :param n:    the size of the windows to chunk into separate periods
-  :type n:    int
-  :param base_chunk: a helper argument for creating all of the staggered chunks, defaults to 0 (original behavior)
-  :type base_chunk: int
-  :return:    pandas DataFrame identical to df with the addition of a chunk column
+    :param df:     data frame of climate data to chunk into different periods
+    :type df:     pandas DataFrame
+    :param n:    the size of the windows to chunk into separate periods
+    :type n:    int
+    :param base_chunk: a helper argument for creating all of the staggered chunks, defaults to 0 (original behavior)
+    :type base_chunk: int
+    :return:    pandas DataFrame identical to df with the addition of a chunk column
 
-  """
+    """
 
     # Check inputs
     df = df.drop_duplicates().reset_index(drop=True).copy()
-    util.check_columns(df, {'year', 'variable', 'value'})
[…]
