+++ .github/workflows/workflow.yml
@@ -4,33 +4,26 @@ on: [push, pull_request]
 
 jobs:
   build:
-      runs-on: ${{ matrix.os }}
-
-      strategy:
-        matrix:
-          os: [ubuntu-latest]
-
-      env:
-        OS: ${{ matrix.os }}
-        PYTHON: '3.9'
-
-      steps:
-
-        - uses: actions/checkout@v3
-
-        - name: Set up Python
-          uses: actions/setup-python@v4
-          with:
-            python-version: 3.9
-
-        - name: Install dependencies
-          run: |
-            python -m pip install --upgrade pip
-            python -m pip install -r requirements.txt
-            python -m pip install .
-            python -c 'import stitches; stitches.install_package_data()'
-        - name: Test and generate coverage report on Linux
-          run: |
-            pip install pytest
-            pip install pytest-cov
-            pytest --cov=./ --cov-report=xml
+    name: Build (${{ matrix.os }}, Python ${{ matrix.python-version }})
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest]
+        python-version: ["3.9"]
+    steps:
+      - uses: actions/checkout@v3
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          python -m pip install -r requirements.txt
+          python -m pip install .
+          python -c 'import stitches; stitches.install_package_data()'
+      - name: Test and generate coverage report on Linux
+        run: |
+          pip install pytest
+          pip install pytest-cov
+          pytest --cov=./ --cov-report=xml

[…]
