+++ .pre-commit-config.yaml
@@ -45,7 +45,7 @@ repos:
         args: ['--line-length=88', '--target-version=py39']
       - id: nbqa-isort
         additional_dependencies: ['isort==5.12.0']
-        args: ['--profile=black', 'py_version=39']
+        args: ['--profile=black']
   - repo: https://github.com/pycqa/pydocstyle
     rev: 6.3.0
     hooks:
@@ -55,7 +55,7 @@ repos:
     hooks:
       - id: blackdoc
         additional_dependencies: ['black==23.3.0']
-        args: ['--line-length=88', '--target-version=py39']
+        args: ['--target-version=py39']
   - repo: https://github.com/adrienverge/yamllint.git
     rev: v1.32.0
     hooks:
