### 2023-07-13 #5525 @observingClouds
> > could you let me know in which aspect you fall under the JOSS COI policy?
>
> In which I aspect I fall under or fail under? I have published papers with both Kalyn and Claudia in the last two years. I am also on a scientific steering committee with Claudia and plan to publish with Kalyn again in the next 12 months.

Thanks @znicholls for giving me additional detail about your COI. Upon consultation with the JOSS team, we will waive the COI for this submission. As you mentioned already, for the future it would be great to know about any potential COI earlier. Maybe we also need to adjust the workflow on our side and ask for the COI before the start of the review process.
