### 2023-08-18 JGCRI/stitches#82 @abigailsnyder
Enhancements and optimization for V1.1 arising from JOSS review
Enhancements:

- [ ] support local archives of data to allow back-compatibility with CMIP5

Optimization:

- [ ] implement old-new testing to speed up optimization type developments without changing behavior
- [ ] utilize `xarray` operations more heavily in the final gridded stitching step to speed up from a few min
- [ ] pass by reference rather than by value where possible
