+++ paper/paper.bib
@@ -54,3 +54,113 @@ DOI = {10.5194/gmd-9-3461-2016}
   url = {https://doi.org/10.1038/nclimate3310},
   doi = {10.1038/nclimate3310}
 }
+
+@article{beusch2020emulating,
+  title={Emulating Earth system model temperatures with MESMER: from global mean temperature trajectories to grid-point-level realizations on land},
+  author={Beusch, Lea and Gudmundsson, Lukas and Seneviratne, Sonia I},
+  journal={Earth System Dynamics},
+  volume={11},
+  number={1},
+  pages={139--159},
+  year={2020},
+  publisher={Copernicus GmbH}
+}
+
+@article{nath2022mesmer,
+  title={MESMER-M: an Earth system model emulator for spatially resolved monthly temperature},
+  author={Nath, Shruti and Lejeune, Quentin and Beusch, Lea and Seneviratne, Sonia I and Schleussner, Carl-Friedrich},
+  journal={Earth System Dynamics},
+  volume={13},
+  number={2},
+  pages={851--877},
+  year={2022},
+  publisher={Copernicus GmbH}
+}
+
+@article{quilcaille2022showcasing,
+  title={Showcasing MESMER-X: Spatially Resolved Emulation of Annual Maximum Temperatures of Earth System Models},
+  author={Quilcaille, Yann and Gudmundsson, Lukas and Beusch, Lea and Hauser, Mathias and Seneviratne, Sonia I},
+  journal={Geophysical Research Letters},
+  volume={49},
+  number={17},
+  pages={e2022GL099012},
+  year={2022},
+  publisher={Wiley Online Library}
+}
+
+@article{hartin2015simple,
+  title={A simple object-oriented and open-source model for scientific and policy analyses of the global climate system--Hector v1. 0},
+  author={Hartin, Corinne A and Patel, Pralit and Schwarber, Adria and Link, Robert P and Bond-Lamberty, BP},
+  journal={Geoscientific Model Development},
+  volume={8},
+  number={4},
+  pages={939--955},
+  year={2015},
+  publisher={Copernicus GmbH}
+}
+
+@article{meinshausen2011emulating,
+  title={Emulating coupled atmosphere-ocean and carbon cycle models with a simpler model, MAGICC6--Part 1: Model description and calibration},
+  author={Meinshausen, Malte and Raper, Sarah CB and Wigley, Tom ML},
+  journal={Atmospheric Chemistry and Physics},
+  volume={11},
+  number={4},
+  pages={1417--1456},
+  year={2011},
+  publisher={Copernicus Publications G{\"o}ttingen, Germany}
+}
+
+@article{smith2018fair,
+  title={FAIR v1. 3: a simple emissions-based impulse response and carbon cycle model},
+  author={Smith, Christopher J and Forster, Piers M and Allen, Myles and Leach, Nicholas and Millar, Richard J and Passerello, Giovanni A and Regayre, Leighton A},
+  journal={Geoscientific Model Development},
+  volume={11},
+  number={6},
+  pages={2273--2297},
+  year={2018},
+  publisher={Copernicus GmbH}
+}
+
+@article{ruane2022climatic,
+  title={The Climatic Impact-Driver Framework for Assessment of Risk-Relevant Climate Information},
+  author={Ruane, Alex C and Vautard, Robert and Ranasinghe, Roshanka and Sillmann, Jana and Coppola, Erika and Arnell, Nigel and Cruz, Faye Abigail and Dessai, Suraje and Iles, Carley E and Islam, AKM Saiful and others},
+  journal={Earth's Future},
+  volume={10},
+  number={11},
+  pages={e2022EF002803},
+  year={2022},
+  publisher={Wiley Online Library}
+}
+
+@article{james2017characterizing,
[…]
