+++ paper/paper.md
@@ -30,85 +30,135 @@ bibliography: paper.bib
 
 # Summary
 
-State of the art impact models characterizing aspects of the interaction between 
-the human and Earth systems require decade-long time series of relatively high
-frequency, spatially resolved and often multiple variables representing 
-climatic impact-drivers. Most commonly these are derived from Earth System
-Model (ESM) output, according to a standard, limited set of future scenarios, 
-the latest being the SSP-RCPs run under CMIP6-ScenarioMIP [@Eyringetal2016;@ONeilletal2016].
-CMIP6/ScenarioMIP is a heavily used product in international
-efforts such as AR6 [citation].
+Understanding the interaction between humans and the Earth system is a 
+computationally daunting task, with many possible approaches depending on 
+resources available and questions of interest. For example, state of the art 
+impact models require decade-long time series of relatively high frequency, 
+spatially resolved and often multiple variables representing climatic impact-
+drivers [@ruane2022climatic]. Most commonly these are derived from the outputs 
+of  detailed, computationally expensive Earth System Models (ESMs) run according 
+to a standard, limited set of future scenarios, the latest being the SSP-RCPs 
+run under CMIP6-ScenarioMIP [@Eyringetal2016;@ONeilletal2016]. At the time of 
+writing, O'Neill et al has been cited more than 1750 times and Eyring et al 
+more than 5000 times, highlighting the broad, general use of this data.
 
-Often, however, impact modeling seeks to explore new scenarios, and/or needs a
-larger set of initial condition ensemble members than are typically available to
-quantify the effects of ESM internal variability. In addition, the recognition that
-the human and Earth systems are fundamentally intertwined, and may feature 
-potentially significant feedback loops, is making integrated, simultaneous modeling 
-of the coupled human-Earth system increasingly necessary, if computationally 
-challenging [@thornton2017biospheric]. 
-For the dual needs of the creation of new scenario realizations and the 
-simplified representation of ESM behavior in a coupled human-Earth system
-modeling framework, climate model output emulators can be the answer. 
-We proposed a new, comprehensive approach to such emulation of CMIP6/ScenarioMIP
-ESM experiments, STITCHES [@tebaldi2022stitches].
-The corresponding `stitches` Python package uses existing archives of ESMs’ 
-scenario experiments to construct new scenarios, or enrich existing initial 
-condition ensembles. Its output has the same characteristics of the ESM output
-emulated: multivariate (spanning potentially all variables that the ESM has 
-saved), spatially resolved (down to the native grid of the ESM), and as high 
-frequency as the original output has been saved at.
+
+Often, however, impact modeling seeks to explore new scenarios that were not 
+part of the ScenarioMIP protocol, and/or needs a larger set of initial condition
+ensemble members than are typically available to quantify the effects of ESM 
+internal variability. In addition, the recognition that the human and Earth 
+systems are fundamentally intertwined, and may feature potentially
+significant feedback loops, is making integrated, simultaneous modeling of 
+the coupled human-Earth system increasingly necessary, if computationally 
+challenging with most existing tools [@thornton2017biospheric]. 
+
+
+For impact modelers, climate model emulators can be the answer to meet both 
+the needs of 1) creating realizations for novel scenarios and 2)achieving a 
+simplified, computationally tractable  representation of ESM behavior in a
+coupled human-Earth system modeling framework.  We proposed a new, 
+comprehensive approach to such emulation of gridded, multivariate ESM 
+outputs for novel scenarios without the computational cost of a full ESM,
+STITCHES [@tebaldi2022stitches]. The approach outlined in Tebaldi et al. should
+be extensible to future CMIP eras, although the `stitches` software at present 
[…]
