+++ paper/paper.md
@@ -68,7 +68,7 @@ is strictly focused on CMIP6/ScenarioMIP data hosted on Pangeo
 The corresponding `stitches` Python package uses existing archives of ESMs’  
 scenario experiments from CMIP6/ScenarioMIP to construct gridded, multivariate 
 realizations of new scenarios provided by reduced complexity climate models 
-(see [@hartin2015simple, @meinshausen2011emulating, @smith2018fair]), or to 
+(see [@hartin2015simple;@meinshausen2011emulating;@smith2018fair]), or to 
 enrich existing initial condition ensembles. Its output has the 
 same characteristics of the ESM output emulated: multivariate (spanning 
 potentially all variables that the ESM has saved), spatially resolved (down to 
@@ -84,21 +84,22 @@ that ESMs require.
 ESM emulation methods generally attempt to preserve the complex statistical
 characteristics of a particular ESM’s outputs for multiple variables and at 
 time scales (often daily or monthly) relevant to impact models. Many 
-existing ESM emulation methods, such as MESMER [@beusch2020emulating, 
-@nath2022mesmer, @quilcaille2022showcasing], rely on 'bottom up' methods, 
+existing ESM emulation methods, such as MESMER 
+[@beusch2020emulating;@nath2022mesmer;@quilcaille2022showcasing], rely 
+on 'bottom up' methods, 
 inferring from the ESM outputs available for training the details of some 
 statistical process (or, more recently, a machine learning algorithm) able to
 generate new realizations with the same spatiotemporal behavior of the original
 ESM outputs, using as input in the generative phase only large scale information,
 like global average temperature (GSAT), that can be generated by a reduced 
-complexity model, such as Hector, Magicc, or FAIR [@hartin2015simple,
-@meinshausen2011emulating, @smith2018fair]
+complexity model, such as Hector, Magicc, or FAIR
+[@hartin2015simple;@meinshausen2011emulating;@smith2018fair]
 
 
 
 The STITCHES approach instead takes a top-down approach inspired by the warming-
 level style of analyses used by past Intergovernmental Panel on Climate Change 
-reports [@SR15, @arias2021climate, @masson2021ipcc, @core2023ipcc]. Specifically,
+reports [@SR15;@arias2021climate;@masson2021ipcc;@core2023ipcc]. Specifically,
 `stitches` takes existing ESM output and intelligently recombines time windows 
 of these gridded, multivariate outputs into new instances of transient, 21st 
 century trajectories by stitching them together on the basis of a target GSAT 
@@ -153,8 +154,9 @@ These outputs from `stitches` can then serve as inputs to impact models with
 little to no code changes in the impact models. It may also be possible to 
 endogenize climate impacts in scenario construction by coupling `stitches` 
 with impact models for multiple sectors and a reduced complexity climate model 
-such as Hector, Magicc, or FAIR [@hartin2015simple, @meinshausen2011emulating, 
-@smith2018fair. With the computational efficiency gains of using emulators,
+such as Hector, Magicc, or FAIR 
+[@hartin2015simple;@meinshausen2011emulating;@smith2018fair. With the 
+computational efficiency of using emulators,
 it may be possible to interactively develop new scenarios with more insight than
 would be possible using multimodel ESM ensemble statistics or using off-the-shelf
 ESM scenarios alone.
