+++ stitches/generate_package_data.py
@@ -4,7 +4,8 @@ import stitches.make_pangeo_table as mk_pangeo
 
 
 
-def generate_pkg_data(smoothing_window=9, chunk_window=9, add_staggered=False):
+def generate_pkg_data(smoothing_window=9, chunk_window=9, add_staggered=False,
+                     anomaly_startYr=1995, anomaly_endYr=2014):
     """ Generate all of the internal package data for stitches, the tas archive,
     matching archive, & the table of pangeo files.
 
@@ -13,7 +14,7 @@ def generate_pkg_data(smoothing_window=9, chunk_window=9, add_staggered=False):
     """
 
     # This takes several hours to run.
-    mk_tas.make_tas_archive()
+    mk_tas.make_tas_archive(anomaly_startYr=anomaly_startYr, anomaly_endYr=anomaly_endYr)
 
     # These two functions run quickly.
     mk_match(smoothing_window=smoothing_window, chunk_window=chunk_window,
