### 2023-08-22 ec862aa
revert script
