### 2023-08-22 221a150
have GSAT time series include units documenting reference period
