### 2023-08-22 294a4a8
add units to GSAT time series and matching_archive package data
