### 2023-08-22 2310404
update recipe so units get carried; error flag tested and works
