### 2023-08-22 1750d70
have units in stitched GSAT time series
