### 2023-09-19 fd3cc49
Some time objects don't have calendars (numpy), but using xarray to ask for calendar will give response independent of the time object being used
