### 2023-09-19 4d1279d
Edited get_netcdf_values() fixed bug when using daily data, and covered one new edge case
