### 2023-09-19 051f830
Edited get_netcdf_values to use xarray to infer time resolution when not supplied
