+++ .github/workflows/workflow.yml
@@ -26,7 +26,7 @@ jobs:
         - name: Install dependencies
           run: |
             python -m pip install --upgrade pip
-            pip install urllib3<2.0
+            python -m pip install urllib3<2.0
             python setup.py install
             python -c 'import stitches; stitches.install_package_data()'
         - name: Test and generate coverage report on Linux
