### 2023-09-21 e419807
missing quote for version spec
