### 2023-09-21 JGCRI/stitches#80 @crvernon
> The tests ran for me although one was failing on main. The instructions are all clear, but using python setup.py is pretty old school now I think (I think there's been a big move to pyproject.toml in order to better support different possible build backends).

We are going to do this in a future release.  We have some logistics at the project level that supports this development that we want to iron out first.  Once we do, we will shift to a more modern approach as you suggest.  Thanks!

> It would be helpful to document how to get coverage information for devs (although it is all based on standard tools so isn't too hard to work out).

Badging in the README on main should now allow for coverage evaluation via codecov on click.  This will allow you to also click through other branches.  Thanks!

> 46% test coverage isn't super high, it could be worth thinking about whether it is feasible to push that figure higher (or remove the less used/tested bits of the API).

Coverage is up to 60.2% in a new dev branch we are working out that will be pulled soon.  This PR will show the change and will make it to main very soon:  https://github.com/JGCRI/stitches/pull/83
