### 2023-09-21 JGCRI/stitches#79 @crvernon
No problem.  We merged the two files and kept the contribution information in the docs and removed the `CONTRIBUTING.md` file from the root dir.  Thanks!

This will go live when the meta PR comes back into main soon.  But the work was done here:  https://github.com/JGCRI/stitches/pull/84
