### 2023-09-21 JGCRI/stitches#77 @crvernon
This has been addressed in a PR here: https://github.com/JGCRI/stitches/pull/85

Also, `stitches.generate_pkg_data` should now be discoverable upon merge of the meta PR.

That PR will make it in with the other review changes in a meta PR shortly.  Thanks!
