### 2023-09-21 JGCRI/stitches#76 @crvernon
Yes, it is unfortunate that the stitches name has already been taken on PyPI.  And also we are currently unable to change the name due to some organizational constraints.  However, we will definitely consider the conda-forge option as I agree this would be great to add.

I am going to close this issue for now but will add this do our development log.  And thank you @Zeitsperre for the willingness to lend a hand!  We will take you up on that!
