+++ stitches/install_pkgdata.py
@@ -22,9 +22,10 @@ class InstallPackageData:
 
     # URL for DOI minted example data hosted on Zenodo
     DATA_VERSION_URLS = {'0.9.1': 'https://zenodo.org/record/7181977/files/data.zip?download=1',
-                         '0.10.0': 'https://zenodo.org/record/7799725/files/data.zip?download=1'}
+                         '0.10.0': 'https://zenodo.org/record/7799725/files/data.zip?download=1',
+                         '0.11.0': 'https://zenodo.org/record/8367628/files/data.zip?download=1'}
 
-    DEFAULT_VERSION = 'https://zenodo.org/record/7167526/files/data.zip?download=1'
+    DEFAULT_VERSION = 'https://zenodo.org/record/8367628/files/data.zip?download=1'
 
     def __init__(self, data_dir=None):
 
