### 2023-09-21 26eee07
add print statments to figure out CI error
