### 2023-09-21 818aae3
try to disable pkg data to speed up tests
