+++ .github/workflows/workflow.yml
@@ -28,7 +28,6 @@ jobs:
             python -m pip install --upgrade pip
             python -m pip install "urllib3<2.0"
             python setup.py install
-            python -c 'import stitches; stitches.install_package_data()'
         - name: Test and generate coverage report on Linux
           run: |
             pip install pytest
