+++ stitches/fx_stitch.py
@@ -53,6 +53,8 @@ def get_netcdf_values(i, dl, rp, fl, name):
     file = rp[name][i]
     start_yr = rp["archive_start_yr"][i]
     end_yr = rp["archive_end_yr"][i]
+    target_start_yr = rp["target_start_yr"][i]
+    target_end_yr = rp["target_end_yr"][i]
 
     # Figure out which index level we are on and then get the
     # xarray from the list.
@@ -66,6 +68,15 @@ def get_netcdf_values(i, dl, rp, fl, name):
     # Time frequency
     freq = xr.infer_freq(times)
 
+    # If daily frequency, check for mis-matched leap-years
+    if ((freq == 'D') | (freq == 'day')):
+        # If using cftime
+        if (type(times) == xr.coding.cftimeindex.CFTimeIndex):
+            target_time_range = xr.cftime_range(start = f'{target_start_yr}-01-01', end = f'{target_end_yr}-12-31', freq='D', calendar=extracted.time.dt.calendar)
+        # Otherwise using pd DatetimeIndex
+        else:
+            target_time_range = pd.date_range(start=f"{target_start_yr}-01-01", end=f"{target_end_yr}-12-31", freq='D')
+
     if type(times) in [xr.coding.cftimeindex.CFTimeIndex, pd.core.indexes.datetimes.DatetimeIndex]:
         yrs = extracted.indexes['time'].year # pull out the year information from the time index
         flags = list(map(lambda x: x in range(start_yr, end_yr+1), yrs))
@@ -75,10 +86,10 @@ def get_netcdf_values(i, dl, rp, fl, name):
     
     # Select only the days within the period
     dat = extracted.sel(time=to_keep)
+    # Output grid as numpy array
+    result = dat[v].values.copy()
 
-    # The number of days we have in our data
-    actual_len = len(dat.time)
-
+    # Get length of time series expected from archive date range
     if ((freq == 'D') | (freq == 'day')):
         # Create series of days using standard daily calendar
         expected_times = pd.date_range(start=str(start_yr) + "-01-01", end=str(end_yr) + "-12-31", freq='D')
@@ -87,17 +98,40 @@ def get_netcdf_values(i, dl, rp, fl, name):
         # If using noleap calendar
         if extracted['time'].dt.calendar == 'noleap':
             # Get number of days over given period with no leap years
-            expected_len = len(expected_times[~((expected_times.month == 2) & (expected_times.day == 29))])
-        # TODO: 360 day calendars not implemented here
+            expected_len = xr.cftime_range(start = f'{start_yr}-01-01', end = f'{end_yr}-12-31', freq='D', calendar='noleap')
+        elif extracted['time'].dt.calendar == '360_day':
+            # Get number of days over given period with 360 day calendar
+            expected_len = xr.cftime_range(start = f'{start_yr}-01-01', end = f'{end_yr}-12-31', freq='D', calendar='360_day')
     else:
         # Number of months over given year range
         expected_len = len(pd.date_range(start=str(start_yr) + "-01-01", end=str(end_yr) + "-12-31", freq='M'))
 
+    # For daily data, target_time_range and expected_len should be made to match
+    if ((freq == 'D') | (freq == 'day')):
+        len_diff = expected_len - len(target_time_range)
+        # If more leap days in archive data, need to remove that number of days (could be at most one?)
+        if(len_diff > 0):
+            # Remove last "len_diff" days
+            result = result[0:(-len_diff),:,:].copy()
+        # If more leap days in target period than archive data, need to add missing days
+        elif(len_diff < 0):
+            # Get shape of data
+            result_shape = result.shape
+            # Extra days shape
+            extra_days_shape = (np.abs(len_diff), result_shape[1], result_shape[2])
+            # Get last "len_diff" days
[…]
