### 2023-10-04 JGCRI/stitches#76 @znicholls
One other thought on this: for the pyam package, the pyam name was already taken on PyPI so it is on PyPI as pyam-iamc instead. Perhaps this could be put on PyPI as stitches-jgcri or something?
