### 2023-10-04 ea62202
Printing before try statement
