### 2023-10-25 #5525 @abigailsnyder
@kthyng my co-author thinks working on this tomorrow and getting the remaining parts done in a week is feasible. The main challenge in their availability has been family caregiving, so it's a little hard to predict.
We _have_ addressed most of the points and I'm happy to post the in-progress R2R document, if that's helpful. Or I'm more than happy to enter this submission into a paused phase if that would work better.

I appreciate the journal's patience and willingness to work with us during a period where more than a few of us have had big life things coming up.
