### 2023-10-25 #5525 @Zeitsperre
All good for me. Life happens.

Some of the comments we made require some significant changes, so I completely understand the need to not rush the implementation.

I had been holding off on opening PRs as the changes might be significant (mainly formatting). If things are on hold, this might be a good time to propose them. Will open a PR if I have some time next week.
