### 2023-10-26 c2133e4
Special naming convention for BASD use-case
