+++ stitches/data/README.md
@@ -1,16 +1,16 @@
-Space holder for package data.
+Space holder for package data. 
 
 To install pre-built package data run:
 
 ```bash
-import stitches
+import stitches 
 stitches.install_package_data()
 ```
 
-For users that would like to generate the run the following:
+For users that would like to generate the run the following: 
 
 ```bash
-import stitches
+import stitches 
 stitches.generate_pkg_data()
 ```
 but note that this will take several hours to run.
