+++ stitches/data/tas-data/README.md
@@ -1,9 +1,9 @@
-Space holder for package data.
+Space holder for package data. 
 
 To install pre built package data run:
 
 ```bash
-import stitches
+import stitches 
 stitches.install_package_data()
 ```
 
@@ -11,6 +11,6 @@ stitches.install_package_data()
 For users that would like to generate the run but notes that this will take several hours to run.
 
 ```bash
-import stitches
+import stitches 
 stitches.generate_pkg_data()
 ```
