+++ stitches/data/temp-data/README.md
@@ -1,6 +1,6 @@
 Space holder for package data. Install the stitches package data by running:
 
 ```bash
-import stitches
+import stitches 
 stitches.install_package_data()
 ```
