### 2023-12-06 fe06c9f
include test data in install
