### 2023-12-06 32c0c4a
add in docs dependencies
