### 2023-12-06 b0656d6
clean join for path
