### 2023-12-06 e3f9550
add in classifiers to setup.py
