### 2023-12-19 a66ab58
make docstrings PEP 257 compliant and consistent
