### 2023-12-19 b30586e
reconfigure paths for tests outside of package
