+++ MANIFEST.in
@@ -4,4 +4,3 @@ include DISCLAIMER
 include notebooks/*.*
 include data/tas-data/README.md
 include data/temp-data/README.md
-include tests/*.csv
