### 2023-12-19 aec9431
add pytest conf file for setup install
