+++ .github/workflows/workflow.yml
@@ -21,7 +21,6 @@ jobs:
           python -m pip install --upgrade pip
           python -m pip install -r requirements.txt
           python -m pip install .
-          python -c 'import stitches; stitches.install_package_data()'
       - name: Test and generate coverage report on Linux
         run: |
           pip install pytest
