+++ tests/test_match.py
@@ -1,4 +1,5 @@
 import unittest
+from importlib import resources
 
 import pandas as pd
 
@@ -20,7 +21,7 @@ class TestMatch(unittest.TestCase):
         operates correctly.
         """
         # Read in some made up target data.
-        path = "test-target_dat.csv"
+        path = resources.files("stitches") / "data" / "example" / "test-target_dat.csv"
         data = pd.read_csv(path)
 
         # Start by checking on a self test & without the drop hist constraint.
@@ -28,7 +29,7 @@ class TestMatch(unittest.TestCase):
         self.assertEqual(data.shape[0], self_match.shape[0])
 
         # Now try matching with the test archive
-        path = "test-archive_dat.csv"
+        path = resources.files("stitches") / "data" / "example" / "test-archive_dat.csv"
         archive = pd.read_csv(path)
         match1 = match_neighborhood(data, archive, tol=0)
         # We know this must be true because the test archive contains duplicates of the
@@ -59,7 +60,7 @@ class TestMatch(unittest.TestCase):
 
         "Test drop_hist_false_duplicates"
         # Read in the match test data.
-        path = "test-match_w_dup.csv"
+        path = resources.files("stitches") / "data" / "example" / "test-match_w_dup.csv"
         match_data = pd.read_csv(path)
         cleaned = drop_hist_false_duplicates(match_data)
         self.assertTrue(nrow(match_data) > nrow(cleaned))
