### 2023-12-19 c3fee47
fix manifest paths
