+++ MANIFEST.in
@@ -2,6 +2,6 @@ include README.md
 include LICENSE
 include DISCLAIMER
 include notebooks/*.*
-include data/tas-data/README.md
-include data/temp-data/README.md
-include data/example/test-*.csv
+include stitches/data/tas-data/README.md
+include stitches/data/temp-data/README.md
+include stitches/data/example/*.csv
