+++ stitches/fx_stitch.py
@@ -453,6 +453,8 @@ def gmat_stitching(rp):
         model = match["archive_model"].unique()[0]
         csv_to_load = [file for file in all_files if (model in file) and (file[0] != ".")][0]
 
+        print(f"HERE:  {csv_to_load}")
+
         data = pd.read_csv(csv_to_load)
 
         # Format the data so that if we have historical years in the future scenarios
