### 2023-12-19 204e559
add comment in contributing about tests
