### 2023-12-20 c0a3e23
fix path to test quickstarter
