+++ stitches/package_data.py
@@ -23,7 +23,6 @@ def fetch_quickstarter_data(variable: str) -> xr.Dataset:
             f"Variable '{variable}' not a valid option.  Choose from 'tas' or 'pr'."
         )
 
-    data_directory = resources.files("stitches"),  "data"
-    filename = f"stitched_CanESM5_{variable_lower}_ssp245~r1i1p1f1~1.nc"
+    f = resources.files("stitches") / "data" / f"stitched_CanESM5_{variable_lower}_ssp245~r1i1p1f1~1.nc"
 
-    return xr.open_dataset(os.path.join(data_directory, filename))
+    return xr.open_dataset(f)
