### 2023-12-20 b7604ab
add progress bar to pangeo fetch
