### 2023-12-20 2665b34
add pre-commit as dev dependency
