### 2023-12-20 47fab7c
flake8 and pydocstring fixes
