### 2023-12-20 #5525 @abigailsnyder
@observingClouds We have addressed all comments and performed a new release of `stitches` (https://github.com/JGCRI/stitches/releases/tag/v0.11.0). All changes in response to the review here are now on `main`.
Attached here is a PDF of our R2R. We pulled every comment we could find  across this review, issues and PRs opened in the `stitches` repo by reviewers into one place to organize and address.

We thank the reviewers and editors for their thorough comments and patience as we have navigated this revision. As it is about to be Winter Holidays, we may be slow to respond in the coming weeks. Hopefully the reviewers and editors will be as well, enjoying a break. We look forward to addressing any additional comments or new reviews in the new year.

[stitches_joss_r2r.pdf](https://github.com/openjournals/joss-reviews/files/13730754/stitches_joss_r2r.pdf)
