### 2024-01-04 #5525 @abigailsnyder
@observingClouds  Thank you for catching this! It was one of our earlier changes we made and it got stranded on a branch. It has been merged into main and can be seen at https://github.com/JGCRI/stitches/blob/main/paper/paper.md now
Thanks again!
Abigail
