### 2024-01-05 #5525 @observingClouds
Dear @Zeitsperre,
Dear @znicholls,
Thank you very much for your review and also sharing your expertise with e.g., setting up the linters and opening some pull requests. Now that your comments have been responded too, I like you to have a last check if everything has been addressed accordingly. In particular, I like you to:
- check all boxes in your checklist if those checkpoints are now satisfied
- check if the code/functions are now well enough documented so that potential new collaborators could extend the functionality to files/scenarios beyond CMIP6, e.g. is it clear how the input data looks like. You both mentioned that this was not clear so far. While I do not expect that this software package currently supports much beyond CMIP6, I do want to make sure that it can be expanded in the future and could be useful during the next CMIPs.
Thank you very much for your time and expertise.

Hauke
