### 2024-01-26 JGCRI/stitches#80 @znicholls
Update as part of next review:

- code coverage is up and all tests now pass
- still using setup.py

I'll close this because it doesn't need to block the review, up to the package maintainers to decide whether they want to follow up or are happy with how they're doing testing.
