### 2024-01-26 JGCRI/stitches#78 @znicholls
Comments from second round of review:

- a script or docs or something that shows how CMIP data is pre-processed, particularly details that are easy to get wrong e.g. area weighting, would be very useful still I think. Yes, this pre-processing is not necessarily part of stitches but it is important to get right in order to use stitches correctly so I think more discussion of this pre-processing step would be worthwhile to give users a better chance of using stitches correctly.

- `install_package_data` seems to download all the data every time, even if data is already downloaded. Is that intended?

- I would put a link to the ESD paper in the notebook where it is mentioned (at the moment it is just referred to but no link is included)
