### 2024-01-26 #5525 @znicholls
@observingClouds I've gone through things again. I think this is now good enough.

My only hesitation was on the functionality question. A key step seems to be pre-processing of CMIP6. In my opinion, this still isn't documented in any real sense. I note that you could argue that this pre-processing is not part of the package, but in order to use it this pre-processed data is key so I would be more comfortable if it were included.

I would also note that the only installation option is still from source. That's fine, it works, but it is a bit odd for a Python package where the barrier to releasing on pypi is so low.

My other comments are in https://github.com/JGCRI/stitches/issues/78, they are not review blocking though.
