### 2024-02-06 #5525 @observingClouds
@abigailsnyder as you can see from the second round of reviews the clarity and reproducibility of the package would strongly benefit from a better description of the input data processing and its format. Please extend the current explanation.

While not necessary, I encourage you to make the package pip installable as znicholls suggests.
