### 2024-02-07 #5525 @abigailsnyder
@observingClouds @znicholls  yes, thank you! I was just waiting for the second reviewer's feedback before responding.

re pip installable - normally this is the standard for our group. Unfortunately for us, there is already a package named stitches on PyPi.  See https://pypi.org/project/stitches/.  So we can't technically publish stitches to PyPi using the stitches name. It's a good reminder for us to check PyPi before naming packages in the future. I am checking with one of my co-authors who knows more about this to see if we can maybe publish the package on PyPi using a name like `jgcri_stitches` for pip installation without have to change much in the package (

I will _happily_ add additional description of input data processing - I agree with Zeb's issue text that it's a tricky thing to do and a  critical part of the package.
> a script or docs or something that shows how CMIP data is pre-processed, particularly details that are easy to get wrong e.g. area weighting, would be very useful still I think. Yes, this pre-processing is not necessarily part of stitches but it is important to get right in order to use stitches correctly so I think more discussion of this pre-processing step would be worthwhile to give users a better chance of using stitches correctly.

Just a few follow up questions so we can do it in a way that is satisfying:
The pre-processing functions are included in the package already so good news is it sounds like this is just a question of us referring to and describing them in the right place so users are aware. Thank you for flagging this! The functions doing the steps referred  primarily in the make_tas_archive.py script. (https://github.com/JGCRI/stitches/blob/main/stitches/make_tas_archive.py) and then they get called simply with `generate_package_data()` (https://github.com/JGCRI/stitches/blob/main/stitches/generate_package_data.py) as one of the steps. This function with its default settings What would be the best way to document these so users are aware of them? I will add  some brief text to the `Install the package data from Zenodo` section of the quickstart, and I'm checking on getting it added to github.io page (https://jgcri.github.io/stitches/reference/api.html) as well. Is there anything else that would be preferable instead of these changes?

RE the other two items in the issue:
> install_package_data seems to download all the data every time, even if data is already downloaded. Is that intended?

>  I would put a link to the ESD paper in the notebook where it is mentioned (at the moment it is just referred to but no link is included)

I will get on both of those today!  The package data is installed every time the function is called, so I will add some text to that section of the quickstart to let users know that, and that they can comment it out if they have already run it once.
