### 2024-02-08 #5525 @znicholls
> if we can maybe publish the package on PyPi using a name like `jgcri_stitches` for pip installation without have to change much in the package

Yep you can do this without issue, see e.g. pyam https://github.com/IAMconsortium/pyam which is under pyam-iamc on PyPI

> What would be the best way to document these so users are aware of them?

Yes that would help. Also just having a notebook called, 'Preparing input data' or something that shows the steps would be very helpful.

> The package data is installed every time the function is called, so I will add some text to that section of the quickstart to let users know that, and that they can comment it out if they have already run it once.

This is a solution. It would be slightly nicer if the downloader just checked if the data was already there before doing the download (trivial using packages like pooch https://www.fatiando.org/pooch/latest/index.html) but it's not a huge issue
