### 2024-02-08 #5525 @abigailsnyder
@znicholls thanks for the suggestions! I'll be working on the Preparing input data notebook and  hopefully while I do that, my colleague will have a few free moments to push things to PyPi (thank you for the example!)
