+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools>=61.0", "wheel", "setuptools-scm"] 
+requires = ["setuptools>=61.0", "wheel"] 
 build-backend = "setuptools.build_meta"
 
 [project]
