+++ setup.py
@@ -3,16 +3,6 @@ import re
 from setuptools import find_packages, setup
 
 
-def readme():
-    """Read and return the content of the README.md file.
-
-    :return: The content of the README.md file.
-    :rtype: str
-    """
-    with open("README.md") as f:
-        return f.read()
-
-
 def requirements():
     """
     Read and parse the 'requirements.txt' file.
@@ -36,10 +26,9 @@ setup(
     packages=find_packages(),
     url="https://github.com/JGCRI/stitches",
     license="BSD 2-Clause",
-    author="Abigail Snyder; Kalyn Dorheim; Claudia Tebaldi",
-    author_email="abigail.snyder@pnnl.gov; kalyn.dorheim@pnnl.gov; claudia.tebaldi@pnnl.gov",
+    author="Abigail Snyder, Kalyn Dorheim, Claudia Tebaldi",
+    author_email="abigail.snyder@pnnl.gov, kalyn.dorheim@pnnl.gov, claudia.tebaldi@pnnl.gov",
     description="Amalgamate existing climate data to create monthly climate variable fields",
-    long_description=readme(),
     python_requires=">=3.9.0",
     include_package_data=True,
     install_requires=requirements(),
