### 2024-02-12 e6f7c1d
fix comma in init
