### 2024-02-12 75f71f3
try reformmating docstring in init
