+++ stitches/__init__.py
@@ -6,6 +6,7 @@ into a single, coherent dataset.
 """
 
 
+from ._version import __version__
 from .fx_match import match_neighborhood
 from .fx_pangeo import fetch_nc, fetch_pangeo_table
 from .fx_recipe import generate_gridded_recipe, make_recipe, permute_stitching_recipes
@@ -16,8 +17,6 @@ from .make_matching_archive import make_matching_archive
 from .make_pangeo_table import make_pangeo_comparison, make_pangeo_table
 from .make_tas_archive import make_tas_archive
 from .package_data import fetch_quickstarter_data
-
-__version__ = "0.11.0"
 
 __all__ = [
     "match_neighborhood",
@@ -35,4 +34,5 @@ __all__ = [
     "make_pangeo_table",
     "make_tas_archive",
     "fetch_quickstarter_data",
+    "__version__",
 ]
