+++ setup.py
@@ -10,6 +10,8 @@ import re
 
 from setuptools import find_packages, setup
 
+from ._version import __version__
+
 
 def requirements():
     """
@@ -24,13 +26,13 @@ def requirements():
         return f.read().split()
 
 
-version = re.search(
-    r"__version__ = ['\"]([^'\"]*)['\"]", open("stitches/__init__.py").read(), re.M
-).group(1)
+# version = re.search(
+#     r"__version__ = ['\"]([^'\"]*)['\"]", open("stitches/_version.py").read(), re.M
+# ).group(1)
 
 setup(
     name="stitches-emulator",
-    version=version,
+    version=__version__,  # version,
     packages=find_packages(),
     url="https://github.com/JGCRI/stitches",
     license="BSD 2-Clause",
