+++ _sources/reference/api.rst.txt
@@ -1,94 +1,0 @@
-.. _pythonapi:
-
-Python API
-=================
-
-**stitches** offers a programmatic API in Python.
-
-.. note::
-
-  For questions or request for support, please reach out to the development team.
-  Your feedback is much appreciated in evolving this API!
-
-
-Core functionality
----------------------------------
-
-stitches.match_neighborhood
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: stitches.match_neighborhood
-
-
-stitches.permute_stitching_recipes
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: stitches.permute_stitching_recipes
-
-
-stitches.generate_gridded_recipe
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: stitches.generate_gridded_recipe
-
-
-stitches.make_recipe
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: stitches.make_recipe
-
-
-stitches.gridded_stitching
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: stitches.gridded_stitching
-
-
-stitches.gmat_stitching
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: stitches.gmat_stitching
-
-
-stitches.fetch_pangeo_table
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: stitches.fetch_pangeo_table
-
-
-stitches.fetch_nc
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: stitches.fetch_nc
-
-
-stitches.make_tas_archive
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: stitches.make_tas_archive
-
-
-
-stitches.make_matching_archive
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: stitches.make_matching_archive
-
-
-stitches.make_pangeo_table
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: stitches.make_pangeo_table
-
-
-stitches.make_pangeo_comparison
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: stitches.make_pangeo_comparison
-
-
-stitches.install_package_data
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: stitches.install_package_data
-
