+++ _sources/getting-started/installation.rst.txt
@@ -1,8 +1,0 @@
-Installation
-============
-
-Install ```stitches``` from GitHub using pip:
-
-.. code-block:: console
-
-  $ pip install git+https://github.com/JGCRI/stitches.git
