+++ .github/workflows/deploy.yml
@@ -1,40 +1,0 @@
-name: Deploy to GitHub Pages
-
-on:
-    push:
-        branches:
-            - main
-
-jobs:
-    build-and-deploy:
-        runs-on: ubuntu-latest
-        steps:
-            - name: Checkout
-              uses: actions/checkout@v2
-            - name: Set up Python
-              uses: actions/setup-python@v2
-              with:
-                  python-version: '3.11'
-
-            - name: Install dependencies
-              run: |
-                python -m pip install --upgrade pip
-                pip install sphinx
-                pip install sphinx-book-theme
-                pip install sphinx-autodoc-typehints
-                pip install sphinx-copybutton
-                pip install sphinx-click
-                pip install nbsphinx
-
-            - name: Build docs
-              run: |
-                make html
-                working-directory: docs
-
-            - name: Deploy to GitHub Pages
-              uses: peaceiris/actions-gh-pages@v3
-              with:
-                  github_token: ${{ secrets.GITHUB_TOKEN }}
-                  publish_dir: ./docs/build/html
-                  publish_branch: gh-pages
-                  keep_files: false
