### 2024-02-12 #5525 @observingClouds
Thanks for the update @abigailsnyder. Could you quickly outline how you addressed these items in https://github.com/JGCRI/stitches/issues/78 and close this issue if appropriate? I just quickly tried to install `stitches-emulator` via pip and it seems like the `requirements.txt` is missing from the [code package hosted to pypi](https://pypi.org/project/stitches-emulator/#files).
