### 2024-02-12 JGCRI/stitches#78 @abigailsnyder
> a script or docs or something that shows how CMIP data is pre-processed, particularly details that are easy to get wrong e.g. area weighting, would be very useful still I think. Yes, this pre-processing is not necessarily part of stitches but it is important to get right in order to use stitches correctly so I think more discussion of this pre-processing step would be worthwhile to give users a better chance of using stitches correctly.
- this has been added https://github.com/JGCRI/stitches/blob/main/notebooks/preparing-input-data.ipynb

> install_package_data seems to download all the data every time, even if data is already downloaded. Is that intended?
- a note that this is the case has been added to the quickstart notebook

> I would put a link to the ESD paper in the notebook where it is mentioned (at the moment it is just referred to but no link is included)
- this has been done
