+++ stitches/install_pkgdata.py
@@ -32,6 +32,7 @@ class InstallPackageData:
         "0.10.0": "https://zenodo.org/record/7799725/files/data.zip?download=1",
         "0.11.0": "https://zenodo.org/records/8367628/files/data.zip?download=1",
         "0.12.0": "https://zenodo.org/records/8367628/files/data.zip?download=1",
+        "0.12.1": "https://zenodo.org/records/8367628/files/data.zip?download=1",
     }
 
     DEFAULT_VERSION = "https://zenodo.org/record/7167526/files/data.zip?download=1"
