### 2024-02-13 #5525 @abigailsnyder
@observingClouds we're still not sure why the distribution is ignoring the requirements.txt file but Chris figured out a work-around and has updated the distribution. I tested on my mac and the installation is successful now.
