### 2024-02-21 JGCRI/stitches#96 @Zeitsperre
[JOSS] Optional QOL improvements
These are a few low effort suggestions to improve the maintainability of the project here.
- Use Dependabot to ensure Python dependencies and the GitHub Actions are regularly kept up-to-date (see: [docs](https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuring-dependabot-version-updates))
- Pull Request and Issue Templates can help with ensuring that relevant information is found in the description (see: [docs](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/creating-a-pull-request-template-for-your-repository); example: [xclim](https://github.com/Ouranosinc/xclim/blob/master/.github/PULL_REQUEST_TEMPLATE.md))
- Code coverage is already being collected in the testing suite. **If you'd like, I can propose a way of sending this to [codecov.io](https://about.codecov.io/) or [coveralls.io](https://coveralls.io).**

Adopting these changes is entirely up to you (has no impact on the JOSS review; the existing changes are good!).
