### 2024-02-21 JGCRI/stitches#97 @Zeitsperre
[JOSS] Contributor Covenant
A very small change that would be good to have is to copy over the `Enforcement` paragraph (see [here](https://www.contributor-covenant.org/version/2/1/code_of_conduct/)) and list a contact email for reporting issues.

This will likely never need to be used, but it's good to tick all the boxes.

Ref: https://github.com/openjournals/joss-reviews/issues/5525
