### 2024-02-21 3df5ac4
grammar adjustments and typo fixes, add software package citations
