+++ paper/paper.bib
@@ -165,3 +165,48 @@ the Sixth Assessment Report of the Intergovernmental Panel on Climate Change},
   year={2023},
   doi={doi: 10.59327/IPCC/AR6-9789291691647.001}
 }
+
+@article{harris2020array,
+  title={Array programming with {NumPy}},
+  author={Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J. van der Walt and Ralf Gommers and Pauli Virtanen and David Cournapeau and Eric Wieser and Julian Taylor and Sebastian Berg and Nathaniel J. Smith and Robert Kern and Matti Picus and Stephan Hoyer and Marten H. van Kerkwijk and Matthew Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and Warren Weckesser and Hameer Abbasi and Christoph Gohlke and Travis E. Oliphant},
+  year={2020},
+  month=sep,
+  journal={Nature},
+  volume={585},
+  number={7825},
+  pages={357--362},
+  doi={10.1038/s41586-020-2649-2},
+  publisher={Springer Science and Business Media {LLC}},
+  url={https://doi.org/10.1038/s41586-020-2649-2}
+}
+
+@article{scikit-learn,
+  title={Scikit-learn: Machine Learning in {P}ython},
+  author={Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V. and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P. and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
+  journal={Journal of Machine Learning Research},
+  volume={12},
+  pages={2825--2830},
+  year={2011}
+}
+
+@software{reback2020pandas,
+  author= {The pandas development team},
+  title={pandas-dev/pandas: Pandas},
+  month=feb,
+  year=2020,
+  publisher={Zenodo},
+  version={latest},
+  doi= {10.5281/zenodo.3509134},
+  url= {https://doi.org/10.5281/zenodo.3509134}
+}
+
+@article{Hoyer_xarray_N-D_labeled_2017,
+  author={Hoyer, Stephan and Joseph, Hamman},
+  doi={10.5334/jors.148},
+  journal={Journal of Open Research Software},
+  month=apr,
+  number={1},
+  title={{xarray: N-D labeled Arrays and Datasets in Python}},
+  volume={5},
+  year={2017}
+}
