### 2024-02-21 #5525 @Zeitsperre
I've done a small pass-through now.

I think it's safe to say that when https://github.com/JGCRI/stitches/pull/97 and https://github.com/JGCRI/stitches/pull/98 are addressed, the majority of my boxes will be ticked. Moving onto the functionality now...
