### 2024-03-28 #5525 @Zeitsperre
Hi @observingClouds I haven't had a chance to check things since my initial testing. It looks as though many of my comments have been turned into issues within the repository, specifically the points concerning refactoring/optimizing calls for `xarray`. If those are going to be addressed later, I imagine that it would be fine to click that box.

I've opened a few other issues and a Pull Request that have yet to be addressed. Once those are closed, I would be fine with checking the remaining issues.
