### 2024-03-28 #5525 @abigailsnyder
@Zeitsperre I am so sorry - I do not get emails from github when an issue or PR gets opened in the `stitches` repo and I'm just totally unused to actively checking.  I will be able to look at these issues and PR next week. Thank you so much for pointing them out here.
