### 2024-04-09 JGCRI/stitches#97 @abigailsnyder
We are in the process of getting something similar in this spirit approved through our laboratory's legal team but for now this is not an option for us as written. We agree it is necessary and  it is reflected in our todo list issue https://github.com/JGCRI/stitches/issues/82#issuecomment-2045008798
