### 2024-04-09 JGCRI/stitches#96 @abigailsnyder
https://github.com/JGCRI/stitches/pull/99 merged
