### 2024-04-09 #5525 @abigailsnyder
@Zeitsperre Thank you! We have responded to all remaining items.
For the contributor covenant https://github.com/JGCRI/stitches/issues/82#issuecomment-2045008798, we agree but we have to have appropriate language approved by our lab.
For the QOL improvements https://github.com/JGCRI/stitches/issues/96, a PR addressing several of these has been reviewed and merged.
For your PR, @crvernon caught one more typo but overall we are happy to merge this. I don't know the extent to which the JOSS process includes a copy-edit pass as other journals do but for the most part it seems like everything from this PR is very much in that space, to me at least?

Thank you so much for your feedback and your patience!
