### 2024-04-29 #5525 @Zeitsperre
@abigailsnyder

Thanks for merging the final PR from me. I'll leave the QOL suggestions issue open on the repository (as there are a few things left to investigate) but I recommend the project for publication now.

Congratulations! Very fascinating library!
