+++ stitches/install_pkgdata.py
@@ -35,6 +35,7 @@ class InstallPackageData:
         "0.12.1": "https://zenodo.org/records/8367628/files/data.zip?download=1",
         "0.12.2": "https://zenodo.org/records/8367628/files/data.zip?download=1",
         "0.12.3": "https://zenodo.org/records/8367628/files/data.zip?download=1",
+        "0.13.0": "https://zenodo.org/records/8367628/files/data.zip?download=1",
     }
 
     DEFAULT_VERSION = "https://zenodo.org/record/7167526/files/data.zip?download=1"
