### 2024-04-30 #5525 @observingClouds
@abigailsnyder thanks for the release and the archive DOI. Could you also check your bibliography and add DOIs to all entries or if none is available provide a url. Thank you.
