+++ paper/paper.bib
@@ -64,7 +64,9 @@ DOI = {10.5194/gmd-9-3461-2016}
   number={1},
   pages={139--159},
   year={2020},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
+  url = {https://doi.org/10.5194/esd-11-139-2020},
+  doi = {10.5194/esd-11-139-2020}
 }
 
 @article{nath2022mesmer,
@@ -75,7 +77,9 @@ DOI = {10.5194/gmd-9-3461-2016}
   number={2},
   pages={851--877},
   year={2022},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
+  url = {https://doi.org/10.5194/esd-13-851-2022},
+  doi = {10.5194/esd-13-851-2022}
 }
 
 @article{quilcaille2022showcasing,
@@ -86,7 +90,9 @@ DOI = {10.5194/gmd-9-3461-2016}
   number={17},
   pages={e2022GL099012},
   year={2022},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  url = {https://doi.org/10.1029/2022GL099012},
+  doi = {10.1029/2022GL099012}
 }
 
 @article{hartin2015simple,
@@ -97,7 +103,9 @@ DOI = {10.5194/gmd-9-3461-2016}
   number={4},
   pages={939--955},
   year={2015},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
+  url = {https://doi.org/10.5194/gmd-8-939-2015},
+  doi = {10.5194/gmd-8-939-2015}
 }
 
 @article{meinshausen2011emulating,
@@ -108,7 +116,9 @@ DOI = {10.5194/gmd-9-3461-2016}
   number={4},
   pages={1417--1456},
   year={2011},
-  publisher={Copernicus Publications G{\"o}ttingen, Germany}
+  publisher={Copernicus Publications G{\"o}ttingen, Germany},
+  url = {https://doi.org/10.5194/acp-11-1417-2011},
+  doi = {10.5194/acp-11-1417-2011}
 }
 
 @article{smith2018fair,
@@ -119,7 +129,9 @@ DOI = {10.5194/gmd-9-3461-2016}
   number={6},
   pages={2273--2297},
   year={2018},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
+  url = {https://doi.org/10.5194/gmd-11-2273-2018},
+  doi = {10.5194/gmd-11-2273-2018}
 }
 
 @article{ruane2022climatic,
@@ -130,7 +142,9 @@ DOI = {10.5194/gmd-9-3461-2016}
   number={11},
   pages={e2022EF002803},
   year={2022},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  url = {https://doi.org/10.1029/2022EF002803},
+  doi = {10.1029/2022EF002803}
 }
 
 @article{james2017characterizing,
@@ -141,20 +155,25 @@ DOI = {10.5194/gmd-9-3461-2016}
   number={2},
   pages={e457},
   year={2017},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  url = {https://doi.org/10.1002/wcc.457},
+  doi = {10.1002/wcc.457}
 }
 
 @article{arias2021climate,
   title={Climate Change 2021: the physical science basis. Contribution of Working Group I to the Sixth Assessment Report of the Intergovernmental Panel on Climate Change; technical summary},
[…]
