+++ paper/paper.bib
@@ -182,7 +182,8 @@ DOI = {10.5194/gmd-9-3461-2016}
 the Sixth Assessment Report of the Intergovernmental Panel on Climate Change},
   author={Core Writing Team, H. Lee and J. Romero (eds.)},
   year={2023},
-  doi={10.59327/IPCC/AR6-9789291691647.001}
+  doi={10.59327/IPCC/AR6-9789291691647.001},
+  url = {https://www.ipcc.ch/report/ar6/syr/downloads/report/IPCC_AR6_SYR_SPM.pdf}
 }
 
 @article{harris2020array,
