### 2024-05-07 #5525 @abigailsnyder
@observingClouds I got the DOIs added in this morning and I have double checked Authors, affiliations, and ORCIDs. I believe @crvernon and I have checked everything else on the author task list and hopefully are good to go
