### 2024-05-07 #5525 @Zeitsperre
I believe you can use the Zenodo DOI for scikit-learn: https://zenodo.org/doi/10.5281/zenodo.591564 (all versions)
