### 2024-05-07 #5525 @Zeitsperre
Additionally, for the IPCC reports, you can refer to this page: https://www.ipcc.ch/report/ar6/wg1/about/how-to-cite-this-report/
