+++ paper/paper.bib
@@ -164,7 +164,8 @@ DOI = {10.5194/gmd-9-3461-2016}
   title={Climate Change 2021: the physical science basis. Contribution of Working Group I to the Sixth Assessment Report of the Intergovernmental Panel on Climate Change; technical summary},
   author={Arias, Paola and Bellouin, Nicolas and Coppola, Erika and Jones, Richard and Krinner, Gerhard and Marotzke, Jochem and Naik, Vaishali and Palmer, Matthew and Plattner, G-K and Rogelj, Joeri and others},
   year={2021},
-  url = {https://www.ipcc.ch/report/ar6/wg1/}
+  url = {https://www.ipcc.ch/report/ar6/wg1/},
+  doi = {10.1017/9781009157896.002}
 }
 
 @article{masson2021ipcc,
@@ -208,7 +209,8 @@ the Sixth Assessment Report of the Intergovernmental Panel on Climate Change},
   volume={12},
   pages={2825--2830},
   year={2011},
-  url = {https://www.jmlr.org/papers/volume12/pedregosa11a/pedregosa11a.pdf}
+  url = {https://www.jmlr.org/papers/volume12/pedregosa11a/pedregosa11a.pdf},
+  doi = {10.5281/zenodo.10951361}
 }
 
 @software{reback2020pandas,
