### 2024-05-08 #5525 @kthyng
Paper:
* 1st paragraph: CMIP6-ScenarioMIP looks strange - is that the normal way to reference this? The rest of the paper you see to use a "/" instead of a "-".
* paragraph 3 pg 2:
* ( see (V Masson): the reference should be inline in the parentheses, otherwise remove the outer parenthesis and keep the reference parenthetical.
* Also comment above ⬆️
