+++ paper/paper.bib
@@ -214,7 +214,7 @@ the Sixth Assessment Report of the Intergovernmental Panel on Climate Change}},
 }
 
 @software{reback2020pandas,
-  author= {The pandas development team},
+  author= {{The {P}andas development team}},
   title={pandas-dev/pandas: {P}andas},
   month=feb,
   year=2020,
