### 2024-05-09 #5525 @abigailsnyder
@kthyng I have addressed the raised issues (and one additional, switching from quotations to backtick on line 75 of the PDF). I have done a proof read on my end and will complete another on the proof editorialbot produces here. Thank you!
