### 2024-05-09 #5525 @kthyng
Congratulations on your new publication @abigailsnyder! Many thanks to @observingClouds and to reviewers @znicholls and @Zeitsperre for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
