### 2024-05-09 #5525 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
